requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
google-generativeai==0.8.0
pdfplumber==0.11.0
reportlab==4.2.2
python-multipart==0.0.12
pydantic==2.9.0
